[#10] Split model classes onto separate files
