+++ app/models/comment_content_interface.ts
@@ -1,0 +1,5 @@
+export interface CommentContentInterface {
+    raw?: string;
+    html?: string;
+    markup?: string;
+}
