+++ app/models/comment_links_interface.ts
@@ -1,0 +1,6 @@
+import {LinkInterface} from "./link_interface";
+
+export interface CommentLinksInterface {
+    self?: LinkInterface;
+    html?: LinkInterface;
+}
