+++ app/models/config_object.ts
@@ -1,0 +1,5 @@
+export class ConfigObject {
+    public static USER: string = 'app:user';
+    public static SOCKET_SERVER: string = 'app:socket_server';
+    public static PULLREQUEST_PROGRESS: string = 'app:pullrequest_progress';
+}
