+++ app/models/event/pull_request_event.ts
@@ -1,0 +1,9 @@
+import {User} from "../user";
+import {PullRequest} from '../pull_request';
+
+export class PullRequestEvent {
+    public actor: User = new User();
+    public sourceEvent: string = '';
+    public pullRequests: PullRequest[] = [];
+    public context: PullRequest = new PullRequest();
+}
