+++ app/models/event/socket_client_event.ts
@@ -1,0 +1,4 @@
+export class SocketClientEvent {
+    public static INTRODUCE: string = 'client:introduce';
+    public static REMIND: string = 'client:remind';
+}
