+++ app/models/event/socket_server_event.ts
@@ -1,0 +1,8 @@
+export class SocketServerEvent {
+    public static PULLREQUESTS_UPDATED: string = 'server:pullrequests:updated';
+    public static PULLREQUEST_UPDATED: string = 'server:pullrequest:updated';
+    public static INTRODUCED: string = 'server:introduced';
+    public static REMIND: string = 'server:remind';
+    public static NEW_COMMENT: string = 'server:comments:new';
+    public static NEW_REPLY_FOR_COMMENT: 'server:comments:reply';
+}
