+++ app/models/link_interface.ts
@@ -1,0 +1,3 @@
+export interface LinkInterface {
+    href: string;
+}
