+++ app/models/notification_interface.ts
@@ -1,0 +1,3 @@
+export interface NotificationInterface {
+    notificationId: string;
+}
