+++ app/models/notification_sound.ts
@@ -1,0 +1,6 @@
+export class NotificationSound {
+    public static NEW_PULLREQUEST: string = 'sounds.new_pull_request';
+    public static APPROVED_PULLREQUEST: string = 'sounds.approved_pull_request';
+    public static MERGED_PULLREQUEST: string = 'sounds.merged_pull_request';
+    public static REMINDER: string = 'sounds.reminder';
+}
