+++ app/models/pull_request_links.ts
@@ -1,0 +1,4 @@
+export class PullRequestLinks {
+    public self: string = '';
+    public html: string = '';
+}
