+++ app/models/pull_request_notification.ts
@@ -1,0 +1,6 @@
+import {NotificationInterface} from './notification_interface';
+
+export class PullRequestNotification implements NotificationInterface {
+    public notificationId: string = '';
+    public pullRequestHtmlLink: string = '';
+}
