+++ app/models/pull_request_progress.ts
@@ -1,0 +1,5 @@
+export class PullRequestProgress {
+    public static PROPORTIONS: string = 'proportions';
+    public static PERCENT: string = 'percent';
+    public static PROGRESS_BAR: string = 'progress_bar';
+}
