+++ app/models/pull_request_state.ts
@@ -1,0 +1,1 @@
+export enum PullRequestState {Open, Merged, Declined}
