+++ app/models/reviewer.ts
@@ -1,0 +1,6 @@
+import {User} from './user';
+
+export class Reviewer {
+    public approved: boolean;
+    public user: User = new User();
+}
