+++ app/models/user.ts
@@ -1,0 +1,4 @@
+export class User {
+    public username: string = '';
+    public displayName: string = '';
+}
